[ci] Split up integration.shard native_assets_test.dart
