+++ packages/flutter_tools/test/integration.shard/isolated/native_assets_agp_version_test.dart
@@ -15,8 +15,6 @@ import '../test_utils.dart' show flutterBin, platform;
 import '../transition_test_utils.dart';
 import 'native_assets_test_utils.dart';
 
-const packageName = 'package_with_native_assets';
-
 /// The AGP versions to run these tests against.
 final agpVersions = <String>['8.4.0'];
